+++ packages/flame/CHANGELOG.md
@@ -7,6 +7,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 Refer to the [CONTRIBUTING guide](https://github.com/lightspeed/flame/blob/master/.github/CONTRIBUTING.md) for more info.
 
+## [Unreleased]
+
+### Fixed
+
+- Add esm builds and their associated esm types to final dist output ([#67](https://github.com/lightspeed/flame/pull/67))
+
 ## 1.2.9 - 2019-12-12
 
 ### Fixed