+++ packages/flame/CHANGELOG.md
@@ -7,6 +7,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 Refer to the [CONTRIBUTING guide](https://github.com/lightspeed/flame/blob/master/.github/CONTRIBUTING.md) for more info.
 
+## [Unreleased]
+
+### Added
+
+- `note`, `receipt-printer`, `hardware`, `usb-plug`, `wifi`, `apple` and `windows` icons ([#81](https://github.com/lightspeed/flame/pull/81))
+
 ## 1.3.0 - 2020-04-20
 
 ### Added